oth: update folder structure
