+++ src/features/addingBots/index.ts
@@ -1,0 +1,1 @@
+export { AddingBots, addingBots } from "./addingBots";
