+++ src/features/help/index.ts
@@ -1,0 +1,1 @@
+export { Help, help } from "./help";
