+++ src/features/language/index.ts
@@ -1,0 +1,1 @@
+export { Language, language } from "./language";
