+++ src/features/profanityFilter/index.ts
@@ -1,0 +1,1 @@
+export { ProfanityFilter, profanityFilter } from "./profanityFilter";
