+++ src/features/settings/index.ts
@@ -1,0 +1,2 @@
+export { Settings, settings } from "./settings";
+export { SettingsAction } from "./settings.types";
