+++ src/features/settings/settings.ts
@@ -7,21 +7,7 @@ import { PAGE_SIZE } from "utils/consts";
 import { isPrismaChatAdmin, prisma, upsertPrismaChat } from "utils/prisma";
 import { getChatHtmlLink, getPagination, getTelegramErrorCode, isCleanCommand } from "utils/telegraf";
 
-export enum SettingsAction {
-  ADDING_BOTS = "cfg-addng-bts",
-  ADDING_BOTS_SAVE = "cfg-addng-bts-sv",
-  CHATS = "cfg-chats",
-  FEATURES = "cfg-ftrs",
-  LANGUAGE = "cfg-lng",
-  LANGUAGE_SAVE = "cfg-lng-sv",
-  PROFANITY_FILTER = "cfg-pf",
-  PROFANITY_FILTER_SAVE = "cfg-pf-sv",
-  REFRESH = "cfg-rfrsh",
-  TIME_ZONE = "cfg-tz",
-  TIME_ZONE_SAVE = "cfg-tz-sv",
-  VOTEBAN = "cfg-vtbn",
-  VOTEBAN_SAVE = "cfg-vtbn-sv",
-}
+import { SettingsAction } from "./settings.types";
 
 export class Settings {
   /**
