+++ src/features/timeZone/index.ts
@@ -1,0 +1,1 @@
+export { TimeZone, timeZone } from "./timeZone";
