+++ src/features/voteban/index.ts
@@ -1,0 +1,2 @@
+export { Voteban, voteban } from "./voteban";
+export { VotebanAction } from "./voteban.types";
