+++ src/features/voteban/voteban.types.ts
@@ -1,0 +1,4 @@
+export enum VotebanAction {
+  BAN = "voteban-ban",
+  NO_BAN = "voteban-no-ban",
+}
